+++ hotdata/models/job_type.py
@@ -30,9 +30,7 @@ class JobType(str, Enum):
     NOOP = 'noop'
     DATA_REFRESH_TABLE = 'data_refresh_table'
     DATA_REFRESH_CONNECTION = 'data_refresh_connection'
-    DATASET_REFRESH = 'dataset_refresh'
     CREATE_INDEX = 'create_index'
-    CREATE_DATASET_INDEX = 'create_dataset_index'
     MANAGED_LOAD = 'managed_load'
 
     @classmethod