+++ test/test_index_entry_response.py
@@ -43,6 +43,7 @@ class TestIndexEntryResponse(unittest.TestCase):
                 index_name = '',
                 index_type = '',
                 metric = '',
+                source_column = '',
                 status = 'ready',
                 updated_at = datetime.datetime.strptime('2013-10-20 19:20:30.00', '%Y-%m-%d %H:%M:%S.%f'),
                 connection_id = '',